Fix corruption of audio in up/down mixing
Pydub does not like non-mono, non-stereo audio segments, and will panic whenever it is told to resample audio into > 2 channels (which was fixed by a previous patch),
AND it also will panic whenever it is told to resample audio that already has > 2 channels. This commit fixes this, at least for the resample method.